updated cleanup to go through subscriptionsByEntityId map values
