+++ application/src/main/java/org/thingsboard/server/controller/plugin/TbWebSocketHandler.java
@@ -541,14 +541,7 @@ public class TbWebSocketHandler extends TextWebSocketHandler implements WebSocke
 
     @Override
     public boolean isOpen(String externalId) {
-        String internalId = externalSessionMap.get(externalId);
-        if (internalId != null) {
-            SessionMetaData sessionMd = getSessionMd(internalId);
-            if (sessionMd != null) {
-                return sessionMd.session.isOpen();
-            }
-        }
-        return false;
+        return externalSessionMap.containsKey(externalId);
     }
 
     private boolean checkLimits(WebSocketSession session, WebSocketSessionRef sessionRef) throws IOException {
